Move BaseViewRestriction model to wagtail.core.models.view_restrictions
